+++ CS411_FINAL_PROJECT/templates/about.html
@@ -1,0 +1,8 @@
+{% extends 'layout.html' %}
+
+{% block body %}
+  <h1>About Us</h1>
+  <p> Sahil, Tatiana, Dennis Nuan</p>
+  <p> This is our group's CS411 project, a way for a user to view all of their social media feeds combined into one. </p>
+  <p> Currently we support twitter and tumblr, but this app can be expanded to include various social feeds. </p>
+{% endblock %}
