+++ CS411_FINAL_PROJECT/templates/following.html
@@ -1,0 +1,10 @@
+{% extends 'layout.html' %}
+
+{% block body %}
+<h1>Following</h1>
+<ul class="list-group">
+  {% for user in follows %}
+    <li class="list-group-item">@{{user}}</li>
+  {% endfor %}
+</ul>
+{% endblock %}
