+++ CS411_FINAL_PROJECT/templates/includes/_navbar.html
@@ -6,16 +6,16 @@
   <div class="collapse navbar-collapse" id="navbarCollapse">
     <ul class="navbar-nav mr-auto">
       <li class="nav-item active">
-        <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
+        <a class="nav-link" href="/home">Home <span class="sr-only">(current)</span></a>
       </li>
       <li class="nav-item">
         <a class="nav-link" href="/about">About</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" href="/articles">Articles</a>
+        <a class="nav-link" href="/settings">Settings</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" href="/twitter">Twitter</a>
+        <a class="nav-link" href="/logout">Logout</a>
       </li>
     </ul>
