+++ CS411_FINAL_PROJECT/templates/layout.html
@@ -3,13 +3,14 @@
   <head>
     <meta charset="utf-8">
     <title>Flask test 1</title>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootswatch/4.1.0/cosmo/bootstrap.min.css">
+    {% block style %}{% endblock %}
   </head>
   <body>
     {% include 'includes/_navbar.html'%}
     <div class="container">
     {% block body%}{% endblock %}
-  </div>
+     </div>
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js">
   </body>
 </html>
