--- CS411_teamassignment3/necessary_imports.txt
@@ -1,4 +1,0 @@
-Hey guys, be sure to install these with pip before running or it won't work.
-
-pip install flask
-pip install python-twitter
