+++ CS411_teamassignment3/templates/about.html
@@ -2,7 +2,5 @@
 
 {% block body %}
   <h1>About Us</h1>
-  <p> Sahil, Tatiana, Dennis Nuan</p>
-  <p> This is our group's CS411 project, a way for a user to view all of their social media feeds combined into one. </p>
-  <p> Currently we support twitter and tumblr, but this app can be expanded to include various social feeds. </p>
+  <p> Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
 {% endblock %}
