+++ CS411_teamassignment3/templates/article.html
@@ -1,0 +1,5 @@
+{% extends 'layout.html' %}
+
+{% block body %}
+  <h1>article number {{id}}</h1>
+{% endblock %}
