+++ CS411_teamassignment3/templates/hometimeline.html
@@ -1,0 +1,10 @@
+{% extends 'layout.html' %}
+
+{% block body %}
+<h1> User's timeline </h1>
+<ul class="list-group">
+  {% for tweet in tweets %}
+    <li class="list-group-item">@{{tweet[1]}}: {{tweet[0]}}</li>
+  {% endfor %}
+</ul>
+{% endblock %}
