+++ CS411_teamassignment3/templates/twitter.html
@@ -1,0 +1,14 @@
+{% extends 'layout.html' %}
+
+{% block body %}
+<h1>Twitter api</h1>
+<form action ="/twitter/echo" method="POST">
+  <input type = "text" name="input"><br>
+  <input type = "submit">
+</form>
+<ul class="list-group">
+  <li class="list-group-item"><a href="twitter/hometimeline"> User (@nhuang54) home timeline of tweets</a></li>
+  <li class="list-group-item"><a href="twitter/usertimeline"> User (@nhuang54)'s tweets</a></li>
+  <li class="list-group-item"><a href="twitter/following"> List of people user is following </a></li>
+</ul>
+{% endblock %}
