--- CS411_teamassignment4/config.cfg
@@ -1,6 +1,0 @@
-CONSUMER_KEY = 'WH3jhSuTMRA3ESj8xInLEsiLe'
-CONSUMER_SECRET = 'c2hPnRpVbv8yudyepiPzZ9ihBbYw6EsnevNDqdi3XnSt3HZH51'
-ACCESS_TOKEN = '976927078489653248-86NxrrQxbrKcdat3Dlxpwaf1aK0euZQ'
-ACCESS_SECRET = '3SP5HnQK4VX9D4OBnttCLXtHjQB5bOrvKY59zhTRHvMM6'
-
-APIKEY = "AIzaSyC9Lgc_qAajV-HzLR0Rjc38ZlZx6Yi4Srg"
